Begin transforming Environment into a protocol.
