+++ AImaKit/Environments.swift
@@ -14,43 +14,17 @@ import Foundation
  * functionality of any environment but must be subclassed to complete the
  * definition and avoid a runtime crash.
  */
-public class AnEnvironment {
+public class AnEnvironment: Environment {
     
-    // STATE
-    
-    /**
-     * The Euclidean space used by this environment.
-     *
-     * For the two-square vacuum world, one may use
-     * ```swift
-     * space = Space(0..<2)
-     * ```
-     * which has two Cartesian coordinates or `Locations`, `[0]` and `[1]`.
-     */
-    let space: Space
-    
-    /**
-     * Except for `Judges`, every object added to the `Environment` goes into
-     * this dictionary along with its `Location`.
-     */
-    var envObjects = Dictionary<EnvironmentObject, Location>()
-    
-    /**
-     * At each time step, each agent in the environment attempts to perform
-     * an action and each judge scores the effort by observing actual changes
-     * to the environment.  This dictionary of dictionaries keeps track of
-     * these associations and scores.
-     */
-    var agentScores = Dictionary<AnAgent, Dictionary<AJudge, Double>>()
-    
-    /**
-     * The views or observers watching or _listening to_ this environment.
-     */
-    var views = Set<EnvironmentView>()
+    //  STATE  ///////////////////////////////////////////////////////////////
 
+    public var space: Space
+    public var envObjects = Dictionary<EnvironmentObject, Location>()
+    public var agentScores = Dictionary<AnAgent, Dictionary<AJudge, Double>>()
+    public var views = Set<EnvironmentView>()
 
-    // INITIALIZATION
-    
+    //  INITIALIZATION  //////////////////////////////////////////////////////
+
     /**
      * Initialize the root environment with a `Space`.
      *
@@ -95,48 +69,6 @@ public class AnEnvironment {
     // CONFIGURATION
     
     /**
-     * Add an object to the environment optionally specifying its location.
-     * If no location is provided, a random one will be chosen.
-     *
-     * - Parameters:
-     *   - thing:  The object to add to environment.
-     *   - location:  Optional location at which to place it.
-     */
-    public func addObject(_ thing: EnvironmentObject, at location: Location? = nil)
-    {
-        if envObjects.keys.contains(thing) {
-            return // There is only one of each Object, the thing is already here.
-        }
-        let position = location ?? space.randomLocation()
-        if let judge = thing as? AJudge
-        {
-            /*
-             * Swift collection iterators like "for (key, value) in myDictionary" are
-             * tricky.  Swift unwraps them for you but then makes them "let constants"
-             * so you cannot modify values this way.  Probably safer, certainly simpler...
-             *
-             * Add new Judge to each Agent's scoring dictionary unless it is already
-             * added.  Judges do not go on "gameboard;" they sit outside the octagon.
-             */
-            for agent in agentScores.keys {
-                if agentScores[agent]![judge] == nil {
-                    agentScores[agent]![judge] = 0.0
-                }
-            }
-        }
-        else
-        {
-            envObjects[thing] = position     // Place everything else on gameboard.
-            if let agent = thing as? AnAgent
-            {
-                envObjects[agent] = position     // Place agent on gameboard.
-                agentScores[agent] = [:]         // Agent has no judges or scores yet.
-                notifyEnvironmentViews(agent);
-            }
-        }
-    }
-    
-    /**
      * Return all environment objects, along with their location, either from a
      * specified location or, if no location is provided, from the entire environment.
      *
@@ -144,21 +76,21 @@ public class AnEnvironment {
      * - Returns: A `Dictionary<EnvironmentObject, Location>` satisfying
      * input criteria.
      */
-    public func getObjects(at location: Location?) -> Dictionary<EnvironmentObject, Location> {
-        var workArea = [EnvironmentObject: Location]() // Start with empty dictionary.
-        if location == nil {
-            workArea = envObjects     // Add all entries.
-        } else {
-            for (key, value) in envObjects {
-                if value == location {
-                    workArea[key] = value // Add only those with matching location.
-                }
-            }
-        }
-        // I think this only makes `result` (the reference) and `Location`s immutable...
-        let result = workArea
-        return result
-    }
+//    public func getObjects(at location: Location?) -> Dictionary<EnvironmentObject, Location> {
+//        var workArea = [EnvironmentObject: Location]() // Start with empty dictionary.
+//        if location == nil {
+//            workArea = envObjects     // Add all entries.
+//        } else {
+//            for (key, value) in envObjects {
+//                if value == location {
+//                    workArea[key] = value // Add only those with matching location.
+//                }
+//            }
+//        }
+//        // I think this only makes `result` (the reference) and `Location`s immutable...
+//        let result = workArea
+//        return result
+//    }
     
     // Not used (yet).
     //
@@ -253,13 +185,13 @@ public class AnEnvironment {
         }
     }
     
-    func notifyEnvironmentViews(_ agent: AnAgent) {
-        for view in views {
-            view.agentAdded(agent, self);
-        }
-    }
-    
-    func notifyEnvironmentViews(_ agent: AnAgent, _ percept: IPercept, _ action: IAction) {
+//    func notifyEnvironmentViews(_ agent: AnAgent) {
+//        for view in views {
+//            view.agentAdded(agent, self);
+//        }
+//    }
+
+    public func notifyEnvironmentViews(_ agent: AnAgent, _ percept: IPercept, _ action: IAction) {
         for view in views {
             view.agentActed(agent, percept, action, self);
         }
@@ -301,7 +233,7 @@ public class EnvironmentView: EnvironmentObject {
      * - Parameter agent: The Agent just added to the Environment.
      * - Parameter source: The Environment to which the agent was added.
      */
-    public func agentAdded(_ agent: AnAgent, _ source: AnEnvironment) {
+    public func agentAdded(_ agent: AnAgent, _ source: Environment) {
     
     }
